main: Move import to own file
